Issue 615 - implementing undo for Find/Replace in Book View This involved a substantial overhaul of the F/R code for this editor, and I changed Replace All to use the Replace Selected code, simply chaining multiple calls together. This is slower, but the sacrifice in speed is probably worth it.
The undo information is stored in jQuery objects rather than being
passed back and forth. These objects are hidden on switching to Code View,
but will still operate when switching back to Book View (unless a glitch
in the focus change system causesit to reload from the DomDocument). The
objects are removed before saving the file to disk.