Switch TempFolder to QTemporaryDir
Refactor TempFolder to use QTemporaryDir internally, instead of
manually creating directories ourselves (which is not race-free).
QTemporaryDir gives us a race-free temporary directory, although still
cleaned manually for performance reasons.

The extra hierarchy (Sigil/scratchpad) in the system temporary
directory is no more used, as the result now is unique enough, and not
shared by all the users on the same machine. As a consequence, there is
no more need to fix the permissions of the scratchpad.