Windows-only fixes for some rare non-ascii bugs
The SIGIL_PREFS_DIR override: getenv() didn't work on Windows when the path specified contained non-ascii characters. Switched to using _wenvget on Windows to remedy.

A hunspell bug prevents the use of dictionary paths that contain non-ascii characters. Prepending a "longpath prefix" to the dictionary paths on Windows forces the use of _wfopen instead of fopen.